+++ ruwatt/src/tensor/operators/sub_tensor_tensor.rs
@@ -43,7 +43,7 @@ impl<T> ops::Sub<Tensor<T>> for Tensor<T> where T: Float {
 mod tests {
     use super::super::super::{Vector, Scalar, Matrix, Tensor};
 
-    fn matrix1234() -> Tensor<f32> {
+    fn matrix1234() -> Tensor<f64> {
         Matrix::square(vec![1.0, 2.0, 3.0, 4.0])
     }
